@@ -16,7 +16,19 @@ load_database_config (struct _SeafileSession *session);
 int
 load_ccnet_database_config (struct _SeafileSession *session);
 
+#ifdef FULL_FEATURE
+#endif
+
+char *
+seaf_gen_notif_server_jwt (const char *repo_id, const char *username);
+
+char *
+seaf_parse_auth_token (const char *auth_token);
+
+void
+split_filename (const char *filename, char **name, char **ext);
+
 int
-load_seahub_database_config (SeafileSession *session);
+seaf_delete_repo_tokens (SeafRepo *repo);
 
 #endif
