@@ -158,7 +158,8 @@ seafile_generate_magic (int version, const char *repo_id,
 }
 
 void
-seafile_generate_pwd_hash (const char *repo_id,
+seafile_generate_pwd_hash (int version,
+                           const char *repo_id,
                            const char *passwd,
                            const char *repo_salt,
                            const char *algo,
@@ -174,7 +175,14 @@ seafile_generate_pwd_hash (const char *repo_id,
      */
     g_string_append_printf (buf, "%s%s", repo_id, passwd);
 
-    pwd_hash_derive_key (buf->str, buf->len, repo_salt, algo, params_str, key);
+    if (version <= 2) {
+        // use fixed repo salt
+        char fixed_salt[64] = {0};
+        rawdata_to_hex(salt, fixed_salt, 8);
+        pwd_hash_derive_key (buf->str, buf->len, fixed_salt, algo, params_str, key);
+    } else {
+        pwd_hash_derive_key (buf->str, buf->len, repo_salt, algo, params_str, key);
+    }
 
     g_string_free (buf, TRUE);
     rawdata_to_hex (key, pwd_hash, 32);
@@ -214,7 +222,8 @@ seafile_verify_repo_passwd (const char *repo_id,
 }
 
 int
-seafile_pwd_hash_verify_repo_passwd (const char *repo_id,
+seafile_pwd_hash_verify_repo_passwd (int version,
+                                     const char *repo_id,
                                      const char *passwd,
                                      const char *repo_salt,
                                      const char *pwd_hash,
@@ -227,7 +236,14 @@ seafile_pwd_hash_verify_repo_passwd (const char *repo_id,
 
     g_string_append_printf (buf, "%s%s", repo_id, passwd);
 
-    pwd_hash_derive_key (buf->str, buf->len, repo_salt, algo, params_str, key);
+    if (version <= 2) {
+        // use fixed repo salt
+        char fixed_salt[64] = {0};
+        rawdata_to_hex(salt, fixed_salt, 8);
+        pwd_hash_derive_key (buf->str, buf->len, fixed_salt, algo, params_str, key);
+    } else {
+        pwd_hash_derive_key (buf->str, buf->len, repo_salt, algo, params_str, key);
+    }
 
     g_string_free (buf, TRUE);
     rawdata_to_hex (key, hex, 32);
