@@ -78,7 +78,8 @@ seafile_generate_magic (int version, const char *repo_id,
                         char *magic);
 
 void
-seafile_generate_pwd_hash (const char *repo_id,
+seafile_generate_pwd_hash (int version,
+                           const char *repo_id,
                            const char *passwd,
                            const char *repo_salt,
                            const char *algo,
@@ -93,7 +94,8 @@ seafile_verify_repo_passwd (const char *repo_id,
                             const char *repo_salt);
 
 int
-seafile_pwd_hash_verify_repo_passwd (const char *repo_id,
+seafile_pwd_hash_verify_repo_passwd (int version,
+                                     const char *repo_id,
                                      const char *passwd,
                                      const char *repo_salt,
                                      const char *pwd_hash,
