@@ -84,11 +84,12 @@ ccnet_user_manager_validate_emailuser (CcnetUserManager *manager,
                                        const char *passwd);
 
 CcnetEmailUser*
-ccnet_user_manager_get_emailuser (CcnetUserManager *manager, const char *email);
+ccnet_user_manager_get_emailuser (CcnetUserManager *manager, const char *email, GError **error);
 
 CcnetEmailUser*
 ccnet_user_manager_get_emailuser_with_import (CcnetUserManager *manager,
-                                              const char *email);
+                                              const char *email,
+                                              GError **error);
 CcnetEmailUser*
 ccnet_user_manager_get_emailuser_by_id (CcnetUserManager *manager, int id);
 
