@@ -1,6 +1,7 @@
 package main
 
 import (
+	"context"
 	"database/sql"
 	"fmt"
 
@@ -63,7 +64,9 @@ func checkQuota(repoID string, delta int64) (int, error) {
 func getUserQuota(user string) (int64, error) {
 	var quota int64
 	sqlStr := "SELECT quota FROM UserQuota WHERE user=?"
-	row := seafileDB.QueryRow(sqlStr, user)
+	ctx, cancel := context.WithTimeout(context.Background(), option.DBOpTimeout)
+	defer cancel()
+	row := seafileDB.QueryRowContext(ctx, sqlStr, user)
 	if err := row.Scan(&quota); err != nil {
 		if err != sql.ErrNoRows {
 			return -1, err
@@ -85,7 +88,9 @@ func getUserUsage(user string) (int64, error) {
 		"owner_id=? AND o.repo_id=RepoSize.repo_id " +
 		"AND v.repo_id IS NULL"
 
-	row := seafileDB.QueryRow(sqlStr, user)
+	ctx, cancel := context.WithTimeout(context.Background(), option.DBOpTimeout)
+	defer cancel()
+	row := seafileDB.QueryRowContext(ctx, sqlStr, user)
 	if err := row.Scan(&usage); err != nil {
 		if err != sql.ErrNoRows {
 			return -1, err
