@@ -21,10 +21,9 @@ struct _SeafileSession {
     char                *tmp_file_dir;
     /* Config that's only loaded on start */
     GKeyFile            *config;
-    GKeyFile            *ccnet_config;
     SeafDB              *db;
     SeafDB              *ccnet_db;
-    SeafDB              *seahub_db;
+    char                *seahub_pk;
 
     SeafBlockManager    *block_mgr;
     SeafFSManager       *fs_mgr;
