@@ -59,7 +59,7 @@ gen_jwt_token ()
 
     jwt_t *jwt = NULL;
 
-    if (!seaf->private_key) {
+    if (!seaf->notif_server_private_key) {
         seaf_warning ("No private key is configured for generating jwt token\n");
         return NULL;
     }
@@ -75,7 +75,7 @@ gen_jwt_token ()
         seaf_warning ("Failed to expire time to jwt\n");
         goto out;
     }
-    ret = jwt_set_alg (jwt, JWT_ALG_HS256, (unsigned char *)seaf->private_key, strlen(seaf->private_key));
+    ret = jwt_set_alg (jwt, JWT_ALG_HS256, (unsigned char *)seaf->notif_server_private_key, strlen(seaf->notif_server_private_key));
     if (ret != 0) {
         seaf_warning ("Failed to set alg\n");
         goto out;
