@@ -121,6 +121,11 @@ static void start_rpc_service (const char *seafile_dir,
                                      searpc_signature_int__string_string_string_string());
 
     searpc_server_register_function ("seafserv-threaded-rpcserver",
+                                     seafile_upgrade_repo_pwd_hash_algorithm,
+                                     "seafile_upgrade_repo_pwd_hash_algorithm",
+                                     searpc_signature_int__string_string_string_string_string());
+
+    searpc_server_register_function ("seafserv-threaded-rpcserver",
                                      seafile_is_repo_owner,
                                      "seafile_is_repo_owner",
                                      searpc_signature_int__string_string());
@@ -199,6 +204,11 @@ static void start_rpc_service (const char *seafile_dir,
                                      seafile_del_file,
                                      "seafile_del_file",
                         searpc_signature_int__string_string_string_string());
+
+    searpc_server_register_function ("seafserv-threaded-rpcserver",
+                                     seafile_batch_del_files,
+                                     "seafile_batch_del_files",
+                        searpc_signature_int__string_string_string());
 
     searpc_server_register_function ("seafserv-threaded-rpcserver",
                                      seafile_copy_file,
@@ -711,6 +721,10 @@ static void start_rpc_service (const char *seafile_dir,
                                      seafile_get_repo_history_limit,
                                      "get_repo_history_limit",
                                      searpc_signature_int__string());
+    searpc_server_register_function ("seafserv-threaded-rpcserver",
+                                     seafile_set_repo_valid_since,
+                                     "set_repo_valid_since",
+                                     searpc_signature_int__string_int64());
 
     /* System default library */
     searpc_server_register_function ("seafserv-threaded-rpcserver",
@@ -1072,6 +1086,8 @@ static void start_rpc_service (const char *seafile_dir,
         exit (1);
     }
 
+    rpc_server->use_epoll = TRUE;
+
     if (searpc_named_pipe_server_start(rpc_server) < 0) {
         seaf_warning ("Failed to start rpc server.\n");
         exit (1);
@@ -1179,10 +1195,7 @@ test_seafile_config(const char *central_config_dir, const char *config_dir, cons
         central_config_dir = ccnet_expand_path (central_config_dir);
     }
 
-    if (seafile_log_init ("-", "debug", "debug") < 0) {
-        fprintf (stderr, "seafile_log_init error: %s\n", strerror(errno));
-        return -1;
-    }
+    seafile_log_init ("-", "debug", "debug", "seaf-server");
 
     srand (time(NULL));
 
@@ -1190,7 +1203,7 @@ test_seafile_config(const char *central_config_dir, const char *config_dir, cons
 
     seaf = seafile_session_new (central_config_dir, seafile_dir, config_dir);
     if (!seaf) {
-        fprintf (stderr, "Error: failed to create ccnet session\n");
+        seaf_error ("Error: failed to create ccnet session\n");
         return -1;
     }
 
@@ -1271,6 +1284,11 @@ main (int argc, char **argv)
         return test_seafile_config (central_config_dir, ccnet_dir, seafile_dir);
     }
 
+    const char *log_to_stdout_env = g_getenv("SEAFILE_LOG_TO_STDOUT");
+    if (g_strcmp0 (log_to_stdout_env, "true") == 0) {
+        daemon_mode = 0;
+    }
+
 #ifndef WIN32
     if (daemon_mode) {
 #ifndef __APPLE__
@@ -1313,7 +1331,7 @@ main (int argc, char **argv)
     if (logfile == NULL)
         logfile = g_build_filename (seafile_dir, "seafile.log", NULL);
 
-    if (seafile_log_init (logfile, "info", "debug") < 0) {
+    if (seafile_log_init (logfile, "info", "debug", "seaf-server") < 0) {
         seaf_warning ("Failed to init log.\n");
         exit (1);
     }
